+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Now you can write `[[test-dependencies]]` in `fixproj.toml`. You can specify dependencies which are only used for tests here. If your project is used as a library, dependencies written in `[[test-dependencies]]` will not be installed. `fix deps add --test <dependency>` adds a dependency to `[[test-dependencies]]`. `fix deps update --test` and `fix deps install --test` will also update lock file and install dependencies including test dependencies.
 
+#### Std
+
+- Add `Std::String::contains` function to check if a string contains a given substring.
+
 ## [1.2.0] - 2026-02-10
 
 ### Added